refactored nav
